+++ CHANGELOG.md
@@ -4,6 +4,8 @@
  - Experiment Overview page no long "hides" chart data for very old experiments. Now it will always select (up to) 600 data points for each line, regardless of the length of the experiment. For larger cluster, this might result in a slower load time - let me know!
  - Initial support for eye-spy.
  - Export option under Inventory page to export the Pioreactors ~/.pioreactor folder. Likewise on the leader.
+ - "sleep" can be used in `pio run pumps ...` to wait between pump actions.
+
 
 #### Bug fixes